+++ library/core/src/ops/place/mod.rs
@@ -128,6 +128,9 @@
 
 use crate::ptr::Pointee;
 
+mod locals;
+mod refs;
+
 /// Marks a type as a place proxy.
 ///
 /// A place proxy "contains" a place (with or without indirection) and provides
@@ -416,28 +419,4 @@ where
     /// Turn a subplace of type `S` into [`Self::Wrapped`].
     #[lang = "wrap_place_wrap"]
     fn wrap(sub: S) -> Self::Wrapped;
-}
-
-#[unstable(feature = "field_projections", issue = "145383")]
-#[lang = "local_handle"]
-pub struct LocalHandle<T: ?Sized> {
-    ptr: NonNull<T>,
-}
-
-impl<T: ?Sized> Clone for LocalHandle<T> {
-    fn clone(&self) -> Self {
-        *self
-    }
-}
-
-impl<T: ?Sized> Copy for LocalHandle<T> {}
-
-impl<T: ?Sized> LocalHandle<T> {
-    pub unsafe fn new(ptr: *mut T) -> Self {
-        Self { ptr: unsafe { NonNull::new_unchecked(ptr) } }
-    }
-
-    pub fn as_ptr(self) -> *mut T {
-        self.ptr.as_ptr()
-    }
 }
